refactor: withdrawal example
