+++ package.json
@@ -48,9 +48,9 @@
     "prettier": "prettier --check .",
     "test": "jest",
     "eg:deposit:ganache": "ts-node -r dotenv/config examples/txDeposit.ts dotenv_config_path=./examples/.env.ganache dotenv_config_debug=true",
-    "eg:transfer:ganache": "ts-node -r dotenv/config examples/txTransfer.ts dotenv_config_path=./examples/.env.ganache dotenv_config_debug=true",
-    "eg:withdrawal": "ts-node examples/withdrawal.ts",
+    "eg:transfer:ganache": "ts-node -r dotenv/config examples/txTransfer.ts dotenv_config_path=./examples/.env.ganache",
+    "eg:withdrawal:ganache": "ts-node -r dotenv/config examples/txWithdrawal.ts dotenv_config_path=./examples/.env.ganache",
     "eg:finalise-withdrawal": "ts-node examples/finaliseWithdrawal.ts",
-    "eg:balances:ganache": "ts-node -r dotenv/config examples/balances.ts dotenv_config_path=./examples/.env.ganache dotenv_config_debug=true"
+    "eg:balances:ganache": "ts-node -r dotenv/config examples/balances.ts dotenv_config_path=./examples/.env.ganache"
   }
 }
